ping,ping6: Do not assume radix point is denoted by '.' (-i option).
In some loacles (e.g., en, ja), radix point is denoted by '.'.
In other locales (e.g., de, fr), radix point is denoted by ','.

The argument parser for interval (-i option) checked '.' to
determine if the argument is integer or not, and used sscanf
with %d or %f.  As we are enabling locale support along with IDN
support, such an assumption is not good at all.

Just let strtod(3) handle difference of locales.

Signed-off-by: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>