arping: Clarify and correct interaction of -c and -w options
The man page is a little unclear, but it seems to say that when -w is not
used, -c causes arping to terminate after sending 'count' packets, but if
-w *is* specified then it terminates only after *receiving* 'count'
packets (or when the timeout expires).  That's not what the code actually
does, or ever did.

This patch clarifies the man page slightly, and corrects the code to
actually implement the descibed behaviour.  This has repeatedly been fixed
incorrectly in Red Hat's packages, causing other problems.  With any luck
this will finally fix it properly.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>