Only emit an audible ping when requested (including flooding).
The changeset 59de58d0f92a2c9b90174e51070aca32e7959e28
"iputils-s20100214" changed flooding to always make an audible ping
when flooding, but didn't provide anyway to disable the audible ping
when flooding.  That's a problem when you want to see the results of
the flood ping, but not hear it.  This patch changes the behavior to
match the normal ping case, no audible ping by default, and -a enables
the ping.

Signed-off-by: David Fries <david@fries.net>