ping: perform dual-stack ping by default
Use libc name resolution properly. The result is that ping is attempted
for all addresses returned by `getaddrinfo()` in order determined by the
C library. That typically means that IPv6 is preferred unless it's
detected as unavailable. This is the expected behavior of client
applications which makes `ping` more attractive for host reachability
tests.

As both IPv4 and IPv6 ping support multiple hops, dual-stack name
resolution is only used for the *last* argument. It is the first one
to be resolved and it determines the address family for all previous
hops.

You can use `-4` to force IPv4 only ping and you can use `ping -6` as an
alias for the good old `ping6`.

Note: The difference between `ping4_run()` and `ping6_run()` prototypes
is there to keep the changes as minimal as possible.