ping: refactor ping socket code
 * Use arguments to affect which sockets are created
 * Use raw sockets only when necessary
 * Handle socket creation errors immediately
 * Use actual socket type for conditions
 * Make `working_recverr` a member of socket_st
 * Pass `sock4` using a pointer

Notes on ping sockets:

With proper setting of `net.ipv4.ping_group_range` sysctl ping now
works without `CAP_NET_RAW` capability unless started with options that
require raw sockets. The default on most systems is to disable the
feature entirely.

It can be fixed by setting the sysctl to allow ping sockets to anyone
or to allow it for a specific group and set that group and the setgid
bit for the ping binary.