Revert "ping_common.c: fix message flood when EPERM is encountered in ping"
This reverts commit a60ad56a18b660ea8158e6ce8d782fb0290329f8.

The reason for this revert is that while the patch above fixed
https://bugzilla.redhat.com/show_bug.cgi?id=1061867 it had weird behavior
with other types of EPERM responses where it could end up in an endless
loop despite the fact that -c flag was specified.  In addition to that the
mentioned bug is also fixed with
commit 8b75f100bbc84aa32a2c35b54c53f48071af6e95
"ping: ICMP error replies while errno < 0 is a hard error"
by Jason A. Donenfeld <Jason@zx2c4.com>
such that this one should no longer be needed anyway.

Signed-off-by: Robert Schiele <rschiele@gmail.com>