arping: Fix ARP protocol field for AX.25 and NETROM
AX.25 and NETROM differ from other, more ethernet-like protocols in that
they are not using a DIX protocol number but the AX.25 PID.  The arping code
doesn't handle this special case resulting in invalid ARP packets being sent.

The interface bpq0 is an AX.25-over-ethernet interface.  Without this
fix:

 # arping -c 1 -I bpq0 172.20.1.3
 ARPING 172.20.1.3 from 172.20.1.2 bpq0
 Sent 1 probes (1 broadcast(s))
 Received 0 response(s)

With this fix:

 # arping -c 1 -I bpq0 172.20.1.3
 ARPING 172.20.1.3 from 172.20.1.2 bpq0
 Unicast reply from 172.20.1.3 [88:98:60:A0:92:40:02]  1.402ms
 Sent 1 probes (1 broadcast(s))
 Received 1 response(s)

Closes: https://github.com/iputils/iputils/pull/360

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
[ pvorel: add new lines for readability ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>