traceroute6: Fix traceroute to IPv6 address
Tracing to an IPv6 address was broken in 40fd68a784 when fixing tracing
to a hostname. It correctly moved the assignment to after the memcpy
into 'to', but that code path is only executed when tracing to a
hostname. The fix just moves the assignment to after the code paths for
filling out 'to' for both IPv6 addresses and hostnames.