Makefile: Initialize LASTTAG variable only for target which actually uses the variable
Before this commit, we always executed

  git describe HEAD

when LASTTAG variable was not set and the Makefile was called which added
a dependency on git (or you would see an error if git command isn't available
on your system).

This commit moves the declaration of the LASTTAG variable into the target's
recipe which is actually using this variable.

Bug: https://bugs.gentoo.org/626994