build-sys: make meson to do everything needed
Perhaps not perfect meson code.  This is my first try with this build
system, so bear with me.

For people who are unfamiliar with meson I added ./configure and Makefile to
get going.  These do bare minimum, as I think we should give hint how to get
going when starting with zero meson knowledge.  Perhaps within years these
become unnecessary.

Notice that idn1 support is removed from configuration options.  It does not
work, so there should not be need to keep it.  Other than that everything
should build as one could expect.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>