arping: support -i interval option
This option will allow to set time interval to use between packets sent.

It may become handy to work around timing issues in the network.  For
example, Linux kernel as of the time of writing by default ignores all ARP
updates that arrive in 1-second interval since previous update for the same
table entry *irrespective of whether the previous update was ignored*.  It
means that calling apring with -c 3 may still not affect Linux network peers
because every consequent update is ignored because of the previous ARP
update (that was also ignored).

In this scenario, we would now be able to call arping -i 2 to work that
kernel behaviour around.

Linux fix can be found in: https://patchwork.ozlabs.org/patch/760372/