ping: do not bind to device when destination IP is on device
Commit f455fee41c077d4b700a473b2f5b3487b8febc1d made ping to bind to device
when a device argument was given.  But that to does not work if destination
IP on device that socket binds to.  See below how to reproduce this issue.

    # ip addr show dev eth0 | grep 'inet '
        inet 192.168.122.100/24 brd 192.168.122.255 scope global eth0
    # ping -I eth0 192.168.122.100
    PING 192.168.122.100 (192.168.122.100) from 192.168.122.100 eth0: 56(84) bytes of data.
    <nothing... loops forever>

Easiest workaround is to cancel bind to device socket option when source and
destination IPs are the same and device option is defined.

Addresses: https://github.com/iputils/iputils/issues/104
Signed-off-by: Sami Kerola <kerolasa@iki.fi>