traceroute6: use monotonic clock
Use clock_gettime(3) CLOCK_MONOTONIC_RAW to determine time difference.  This
measurement is ideal to this use case, as in traceroute6 one does not care
what is the time but how long packages are traveling back and forth.  Notice
that CLOCK_MONOTONIC_RAW is not subject to discontinuous jumps by system
administrator manually adjusting time, or incremental adjustments performed
by adjtime(3) and NTP.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>