build-sys: change build defaults
Recently added localization support should needs to be turned on by default
to gain traction from translations.  Without this it is hopeless to get
contributions going on.

The rarpd is obsoleted by BOOTP and later DHCP protocols.  There should not
be reason why this executable ought to be build and packaged by default to
pretty much all distributions.

The traceroute6 is not needed, tracepath -6 can perform the same function,
so drop it from default build as well.

Addresses: https://github.com/iputils/iputils/issues/162
CC: Petr Vorel <petr.vorel@gmail.com>
Signed-off-by: Sami Kerola <kerolasa@iki.fi>