tftpd: make nak() data handling concise
Former strcpy() was not great choise when copying to void pointer,
memcpy() is more apt.  This function change fixes warning that fortify
is complaining, see reference commit for details.

While on it lets use ssize_t type to hold a value that refers to a size,
with explanation where the value 5 originates (it is the header).

Reference: 86323ed69b8cf227364173584e76167974eef2e6
Signed-off-by: Sami Kerola <kerolasa@iki.fi>