ping: reindent source files
This change aims to improve code readability.  Cost of the change is old
diff merge conflicts.  So I need to be fair and promise to help sorting them
out.  Assuming distributions have patches on top of iputils this change
broke it would be best to get such diffs to upstream.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>