common: use single ARRAY_SIZE definition across files
The ARRAY_SIZE is Rusty Russell’s version, that looks a bit more complicated
than the original, but there is a reason to that.  This macro will ensure
macro is used properly, and fail compilation if it is not.  See reference
link for further explanation.

Reference: http://zubplot.blogspot.com/2015/01/gcc-is-wonderful-better-arraysize-macro.html
Signed-off-by: Sami Kerola <kerolasa@iki.fi>