doc: Use namespace correctly
The files declared xmlns:db but did not use the db namespace at all.
They did not define the default namespace at all, which coincidentally
worked with Docbook 4 stylesheets, making them think the files were
written in Docbook 4.

I fixed the namespaces of the documents and switched to the correct
Docbook 5 stylesheets.

Signed-off-by: Jan Tojnar <jtojnar@gmail.com>