build-sys/doc: make xsltproc required
Building of all man pages is dependent on xsltproc, so make it a hard
requirement. Also, don't continue the configuration phase and error out
when the building of the man pages was required but xsltproc doesn't
work. Otherwise, the step of building the man pages is silently skipped
even though it was required.