ping: allow user to specify VRF and source IP
Without this, the options for sending a ping in the context of a VRF are
limited.

We can send a ping with a specific source IP address.  For example:

  ping 10.1.1.3 -I 10.1.1.2

We can send a ping in the context of a Linux VRF.  For example:

  ping 10.1.1.3 -I vrf_red

However, when pinging in the context of a VRF, Linux does not always choose
a sensible source IP address – the source IP might not belong to the VRF.
As a result, the ping won’t get a response.  As a result, we want to be able
to specify both a VRF and a source IP address when initiating a ping.  For
example:

  ping 10.1.1.3 -I vrf_red -I 10.1.1.2

Ping reads in the command line parameters fine and sets up the 'source' and
'device' variables, but currently ignores the device if the source IP
address is non-zero.  This commit adds a branch to ping.c that does the
socket bind to the device even in the case where the source IP is non-zero.
This branch is based on the existing case where source IP is zero, but
simplified a bit because we've already got a source IP address to use.