arping, tracepath, traceroute6: use clock_gettime(CLOCK_MONOTONIC)
The CLOCK_MONOTONIC_RAW does not get ntp adjustments, that can be considered
as a good thing when trying to avoid time jumps.   But downside of RAW time
is that it does not get adjtime(3) adjustments resulting to incorrect lenght
of a time period.   After all ntp has two functions, it tries to get clock
to display correct time and make it run right pace.   The later is what is
what is needed when measuring time differences, so CLOCK_MONOTONIC is the
indeed better choice.

Sure enough an unlucky measurement might to badly wrong when in between
start and end of a latency measurement adjtime(3) does an adjustment.   Due
to nature of this problem these freak measurements are hard to reproduce,
and hopefully rather rare.   If this really turns out to be a problem then
the question should be what sort of incorrect time measurement is preferred.

Reference: https://github.com/iputils/iputils/pull/213
CC: Petr Vorel <petr.vorel@gmail.com>
Signed-off-by: Sami Kerola <kerolasa@iki.fi>