clockdiff: add --time-format option
As mentioned in manual page iso time format has timezone and it is easy to
parse.  Petr mentioned in pull request it would be better not to change
default output format, that lead me to make time formating to depend on
option argument.

Reference: https://github.com/iputils/iputils/pull/231
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Sami Kerola <kerolasa@iki.fi>