ping: fix ping6 name resolution for hostnames with link-local addresses
getaddrinfo has several potential interfaces it can use to fulfill hostname
requests with the nss, namely nss_gethostbyname{,2,3,4}_r. It selects
which interface to choose based on the passed hints. Unfortunately, the
only interface with the potential to indicate an appropriate scope id
for link-local IPv6 addresses is nss_gethostbyname4_r, which is
necessarily dual-stack. As a result, getaddrinfo avoids that interface
when the requested address family is not AF_UNSPEC causing ping6
requests to fail for hostnames that resolve to link-local addresses
because of an unspecified scope id, even when they could succeed in the
dual-stack case.

This is a problem when using a local resolver such as systemd-resolved
that synthesizes resource records in some cases containing link-local
addresses. Although `ping -6 _gateway` seems like an appropriate test
for IPv6 connectivity, it fails for the above reason.

This commit causes ping to always request the dual-stack name resolution
behavior, even when it intends only to ping IPv6 hosts. If name
resolution returns an addrinfo list that does not contain an address
from the appropriate family, it intentionally fails with EAI_ADDRFAMILY.