traceroute6: add handling for source route header
This patch adds handling for source header. After adding in the nexthdr
evaluation chain it begun to work in my case to run traceroute6 in a
IPv6 source route network.

Note there existing generic extension header which have all mostly the
same header setup as type and length in 8 byte octects... sadly it
doesn't fit for every case. May not the best soltution but it fix my case.