ping: use struct timespec to measure time
Use the struct timespec to measure how much is the time, and various related
intervals.  Aim of this change is to be exact how time calculation happen,
so that it as easy as possible to understand what happens with them.

The exponentially weighted moving average changed slightly while writing
this change.  Earlier implementation was so hard to follow that after this
change maths may have changed, but even if they did at least now one should
be able to understand the equation - and fix if needed.

Notice that there is a need for follow up to remove 'next' millisecond int
variable in pinger() and schedule_exit().  That change will be complicated
enough to be separate commit.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>