arpping: make update neighbours work again
The arping is using inconsistent sender_ip_addr and target_ip_addr in
messages.  This causes the client receiving the arp message not to update
the arp table entries.

The specific performance is as follows:

There is a machine 2 with IP 10.20.30.3 configured on eth0:0 that is in the
same IP subnet as eth0.  This IP was originally used on another machine 1,
and th IP needs to be changed back to the machine 1.  When using the arping
command to announce what ethernet address has IP 10.20.30.3, the arp table
on machine 3 is not updated.

Machine 3 original arp table:

    10.20.30.3  machine 2 eth0:0    00:00:00:00:00:02
    10.20.30.2  machine 2 eth0      00:00:00:00:00:02
    10.20.30.1  machine 1 eth0      00:00:00:00:00:01

Create interface eth0:0 on machine 1, and use the arping command to send arp
packets.  Expected outcome on machine 3:

    10.20.30.3  machine 1 eth0:0    00:00:00:00:00:01
    10.20.30.2  machine 2 eth0      00:00:00:00:00:02
    10.20.30.1  machine 1 eth0      00:00:00:00:00:01

Actual results on machine 3:

    10.20.30.3  machine 2 eth0:0    00:00:00:00:00:02
    10.20.30.2  machine 2 eth0      00:00:00:00:00:02
    10.20.30.1  machine 1 eth0      00:00:00:00:00:01

Fixes: https://github.com/iputils/iputils/issues/298
Fixes: 68f12fc4a0dbef4ae4c404da24040d22c5a14339
Signed-off-by: Aichun Li <liaichun@huawei.com>