ping: process interrupts in ping*_receive_error_msg
If recvmsg(..., MSG_ERRQUEUE) returned error, ping can misinterpret
EHOSTUNREACH from recvmsg in the main loop as error that happened
from his own packets and print the following message:

    ping: recvmsg: No route to host

However, this also could be an error from other ping, that was received
by this process because of RAW socket. So it's better to skip this error
and process it properly in the next cycle.

Signed-off-by: Artem Leshchev <matshch@yandex-team.ru>