ping: Set mark (SO_MARK) on probe socket
If the bind (source) IP is is not explicitly specified with "-I" then
ping creates a "probe" socket to determine the source IP.  However, if a
mark is configured with "-m", the current code does not apply this mark
to the "probe" socket.  If policy routing is configured such that the
mark affects source IP selection, then the wrong source IP will be used.

In addition, if policy routing is configured such that the destination
will be unreachable unless the mark is set, then the probe socket will
fail and ping will exit with an error.

Therefore set the mark on the probe socket in addition to the actual
ping socket.

Testing emulates the real scenario requires to disable rp_filter (source
validation), use 2 netns to avoid mangling sysctl rp_filter setup:

    IPUTILS_NS1="ip netns exec iputils_ns1"
    IPUTILS_NS2="ip netns exec iputils_ns2"
    ip netns add iputils_ns1
    ip netns add iputils_ns2
    ip link add name iputils_veth1 type veth peer name iputils_veth2
    ip link set dev iputils_veth1 netns iputils_ns1
    ip link set dev iputils_veth2 netns iputils_ns2
    $IPUTILS_NS1 ip link set dev iputils_veth1 up
    $IPUTILS_NS1 ip addr add 10.0.0.2/24 dev iputils_veth1
    $IPUTILS_NS1 ip rule add priority 101 fwmark 0x1/0x1 table 1
    $IPUTILS_NS1 ip route add table 1 10.0.1.0/24 dev iputils_veth1 proto kernel scope link src 10.0.0.2
    $IPUTILS_NS1 sysctl -w net.ipv4.conf.all.rp_filter=0
    $IPUTILS_NS1 sysctl -w net.ipv4.conf.iputils_veth1.rp_filter=0
    $IPUTILS_NS2 ip link set dev iputils_veth2 up
    $IPUTILS_NS2 ip addr add 10.0.0.1/24 dev iputils_veth2
    $IPUTILS_NS2 ip addr add 10.0.1.1/24 dev iputils_veth2

    $IPUTILS_NS1 iputils.patched/builddir/ping/ping 10.0.1.1 -c1 -m1

    ip netns del iputils_ns1
    ip netns del iputils_ns2

Closes: https://github.com/iputils/iputils/pull/340

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Paul Donohue <git@PaulSD.com>
[ pvorel: use unsigned int as mark, move opt_mark check outside of function,
add warning SO_MARK not supported ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>