ping: Set mark (SO_MARK) on probe socket
If the bind (source) IP is is not explicitly specified with "-I" then
ping creates a "probe" socket to determine the source IP.  However, if a
mark is configured with "-m", the current code does not apply this mark
to the "probe" socket.  If policy routing is configured such that the
destination will be unreachable unless the mark is set, then the probe
socket will fail and ping will exit with an error.

Therefore set the mark on the probe socket in addition to the actual
ping socket.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Paul Donohue <git@PaulSD.com>
[ pvorel: use unsigned int as mark, move opt_mark check outside of function,
add warning SO_MARK not supported ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>