Ping: Remove binding of ICMP socket
When specifying an interface in the ping command, if trying to ping an
address that is not reachable then no errors would be received. This was
because the ICMP socket was being bound to the source interface but the
errors were coming through the loopback interface.
To fix this the ICMP socket should not be bound to the interface.
Instead IPv4 ping should copy IPv6's use of cmsg. A cmsg header will
contain information about the interface and will then be included in the
msg header when sending a ping. This means the outgoing ping uses the
correct interface but incoming messages from other interfaces can still
be received on the socket.