rdisc: Add support for capability/setuid
via SETCAP_OR_SUID_RDISC variable.

NOTE: This is strongly discouraged as normal uses could possibly modify
routing tables, recommended way is to run it as systemd service which
does not require it to be enabled. It's added just for non-systemd
users, which should run allow it to run just for dedicated user.

Closes: #347, #348

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Artur Lebedev <ras_atari@mail.ru>
[ Default off, add explanation, fix configure output, rewrite commit
message, join commits. ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>