tests: Don't fail when ip binary missing
Failure to run `ip' should produce the same results as running it and
discovering that, indeed, IPv6 is not available. But meson aborted
because a missing program is a stricter issue than one that simply
returns non-zero. So, do an upfront check for the program and disable
those tests if it cannot be found.

While we are at it, optimize the function call chain to drop an unneeded
.strip() before .contains()

Fixes: https://github.com/iputils/iputils/issues/383
Closes: https://github.com/iputils/iputils/pull/390

Reviewed-by: Noah Meyerhans <noahm@debian.org>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
[ pvorel: warn about missing ip binary ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>