tests: do not error out if ip program cannot be found
Failure to run `ip` should produce the same results as running it and
discovering that, indeed, ipv6 is not available. But meson aborted
because a missing program is a stricter issue than one that simply
returns non-zero. So, do an upfront check for the program and disable
those tests if it cannot be found.

While we are at it, optimize the function call chain to drop an unneeded
.strip() before .contains()

Fixes #383

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>