man/ping: Add "IPV6 LINK-LOCAL DESTINATIONS" section
ping does not always work when ping IPv6 site-local with -I option
instead of '%'-notation if ICMP datagram socket is used (more common
nowadays, it works as expected on RAW socket).

It was fixed on kernel 5.17:

    2afc3b5a31f9 ("ping: fix the sk_bound_dev_if match in ping_lookup")
    35a79e64de29 ("ping: fix the dif and sdif check in ping_lookup")
    cd33bdcbead8 ("ping: remove pr_err from ping_lookup")

and backported on stable/LTS kernels: 5.15.19, 5.10.96, 5.4.176,
4.19.228, 4.14.265. It was not realistic to fix 4.9 LTS branch,
thus kernels from 3.0 to 4.13 are unfixed.

It works only on glibc (at least musl does not support it).

It also requires ping >= s20180629.

Link: https://github.com/iputils/iputils/issues/398
Link: https://github.com/iputils/iputils/pull/405

Co-developed-by: Benjamin Poirier <benjamin.poirier@gmail.com>
[ pvorel: Benjamin is the author of the text in the man page ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>