ping: Add support for Identifier field for icmp ECHO_REQUEST
Identifier field can be set with SOCK_DGRAM if it's not 0.
If someone wants to set identifier field to 0 this person need to use
RAW_SOCKETS. When using SOCK_DGRAM the identification field is set in when
binding to an address with port set to the value of desired identier field.
net/ipv4/af_inet.c:__inet_bind function at some point calls
net/ipv4/ping.c:ping_get_port where value of the port is assigned to
isk->inet_num which is later used fror identifier. When passing 0
to the ping_get_port function it goes to "default path", which is
incrementing global value of identifier field by one and using it.
That's the reason of using RAW_SOCKETS when needing to set identifier
to 0.