ping: Refactor style in create_socket()
There will be more code in socket error handling, thus return on valid
socket to reduce indent level.

sock->socktype can be assigned, because all checks for validity are done
via sock->fd.

NOTE: check for 0 in str_socktype() is still needed for -4 or -6,
because only socket for chosen IP version is being created, thus
socktype of the other is 0.

Signed-off-by: Petr Vorel <pvorel@suse.cz>