ping: Fix the errno handling for strtod
The setlocale(LC_ALL, "") following the strtod() for the option 'i'
can fail if the LC_CTYPE is invalid.
Hence the errno check following the setlocale(LC_ALL, "") thinks
wrongly that strtod() failed with the errno and prints a warning.
The errno got from the execution of strtod() is saved and restored
after setlocale() to be checked for any errors.

Co-Developed-by: Sriram Rajagopalan <sriramr@arista.com>
Signed-off-by: Sriram Rajagopalan <sriramr@arista.com>
Signed-off-by: Jacek Tomasiak <jtomasiak@arista.com>