Update Debian CI environments
Remove configuration that assumed that "oldstable" referred to Debian 9,
as this hasn't been the case since the Debian 11 release in
mid-2021 and Debian 9 is fully EOL at this point anyway.

That fixes error in CI:
E: The repository 'http://http.debian.net/debian stretch-backports Release' does not have a Release file.

Closes: https://github.com/iputils/iputils/pull/462
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Noah Meyerhans <noahm@debian.org>
[ pvorel: keep generic distro names ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>