ping: Remove check for invalid ident
When the ident is -1 the ident for the group of requests is chosen by
the kernel and most likely it's not 65535. That's why it was not printed
in 43e38f2. But since previous commit random ident was set even for ICMP
datagram socket and we don't allow to set ident < 0. Thus it should be
safe to always print ident.

Link: https://github.com/iputils/iputils/pull/490#issuecomment-1730353964
Link: https://github.com/iputils/iputils/issues/489#issuecomment-1728967782
Reported-by: Miloslav Hůla
Signed-off-by: Petr Vorel <pvorel@suse.cz>