ping: More precise check for our packet
Ping run with Nagios via Debian specific check-host-alive Nagios plugin
(e.g. ping -n -v -D -W 1 -i 1 -c 5 -M 'do' -s 56 -O "$Host")
can run many ping in parallel (e.g. 75-100 in the reported issue).

This leads to the collisions, because ICMP identification field is only
16 bits (65535). The problem is not new, but it was hidden until
5f6bec5 ("ping: Print reply with wrong source with warning")

We consider this warning useful, therefore instead of reverting another
check for sequence was added.

Other possible solutions to this:

* Revert 5026c22 ("ping: use random value for the identifier field")
Ping in the past used PID as ICMP identification field. This was
considered insecure, thus it was changed to use pseudo random number in
5026c22 ("ping: use random value for the identifier field")
although there is no CVE and the same approach is used in fping and
busybox's ping, thus it's questionable if it was really needed.

* Add another check in the ICMP optional data
We could add 128 bit random value to check. But we already use struct
timeval if packet size is big enough for it (>= 16 bits), therefore we
could use it for comparing for most of the packet sizes (the default is
56 bits).

Fixes: https://github.com/iputils/iputils/issues/489
Reported-by: Miloslav Hůla <miloslav.hula@gmail.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>